+++ NEWS.md
@@ -1,5 +1,6 @@
 ## pyobsplot (development version)
 
+- Autocompletion of Plot methods should now be working in IDEs.
 
 
 ## pyobsplot 0.3.2